+++ tests/SoCreate.ServiceFabric.PubSub.Tests/GivenDefaultBrokerEventsManager.cs
@@ -12,6 +12,96 @@ namespace SoCreate.ServiceFabric.PubSub.Tests
     public class GivenDefaultBrokerEventsManager
     {
         [TestMethod]
+        public async Task WhenSubscribedEventsAreAdded_ThenCallbacksAreExecuted()
+        {
+            var count = 0;
+            var manager = new DefaultBrokerEventsManager();
+            manager.Subscribed += (queueName, subscriber, messageType) =>
+            {
+                count++;
+                return Task.CompletedTask;
+            };
+            await manager.OnSubscribedAsync("myqueue", new ServiceReferenceWrapper(new ServiceReference()), "myMessageType");
+
+            Assert.AreEqual(1, count);
+        }
+
+        [TestMethod]
+        public async Task WhenUnsubscribedEventsAreAdded_ThenCallbacksAreExecuted()
+        {
+            var count = 0;
+            var manager = new DefaultBrokerEventsManager();
+            manager.Unsubscribed += (queueName, subscriber, messageType) =>
+            {
+                count++;
+                return Task.CompletedTask;
+            };
+            await manager.OnUnsubscribedAsync("myqueue", new ServiceReferenceWrapper(new ServiceReference()), "myMessageType");
+
+            Assert.AreEqual(1, count);
+        }
+
+        [TestMethod]
+        public async Task WhenMessagePublishedEventsAreAdded_ThenCallbacksAreExecuted()
+        {
+            var count = 0;
+            var manager = new DefaultBrokerEventsManager();
+            manager.MessagePublished += message =>
+            {
+                count++;
+                return Task.CompletedTask;
+            };
+            await manager.OnMessagePublishedAsync(new MessageWrapper());
+
+            Assert.AreEqual(1, count);
+        }
+
+        [TestMethod]
+        public async Task WhenMessageQueuedToSubscriberEventsAreAdded_ThenCallbacksAreExecuted()
+        {
+            var count = 0;
+            var manager = new DefaultBrokerEventsManager();
+            manager.MessageQueuedToSubscriber += (queueName, subscriber, messageType) =>
+            {
+                count++;
+                return Task.CompletedTask;
+            };
+            await manager.OnMessageQueuedToSubscriberAsync("myqueue", new ServiceReferenceWrapper(new ServiceReference()), new MessageWrapper());
+
+            Assert.AreEqual(1, count);
+        }
+
+        [TestMethod]
+        public async Task WhenMessageDeliveredEventsAreAdded_ThenCallbacksAreExecuted()
+        {
+            var count = 0;
+            var manager = new DefaultBrokerEventsManager();
+            manager.MessageDelivered += (queueName, subscriber, messageType) =>
+            {
+                count++;
+                return Task.CompletedTask;
+            };
+            await manager.OnMessageDeliveredAsync("myqueue", new ServiceReferenceWrapper(new ServiceReference()), new MessageWrapper());
+
+            Assert.AreEqual(1, count);
+        }
+
+        [TestMethod]
+        public async Task WhenMessageDeliveryFailedEventsAreAdded_ThenCallbacksAreExecuted()
+        {
+            var count = 0;
+            var manager = new DefaultBrokerEventsManager();
+            manager.MessageDeliveryFailed += (queueName, subscriber, messageType, exception) =>
+            {
+                count++;
+                return Task.CompletedTask;
+            };
+            await manager.OnMessageDeliveryFailedAsync("myqueue", new ServiceReferenceWrapper(new ServiceReference()), new MessageWrapper(), new Exception("Failed to deliver"), 10);
+
+            Assert.AreEqual(1, count);
+        }
+
+        [TestMethod]
         public void WhenInsertingConcurrently_ThenAllCallbacksAreMadeCorrectly()
         {
             bool hasCrashed = false;
@@ -54,7 +144,7 @@ namespace SoCreate.ServiceFabric.PubSub.Tests
                     }
                 }, i);
 
-                
+
             }
 
             mr.Set();